requirements.txt
certifi==2022.5.18.1
cffi==1.15.0
click==8.1.3
dataclasses==0.8
future==0.18.2
numpy==1.22.1
olefile==0.46
pathspec==0.9.0
Pillow==7.2.0
platformdirs==2.5.1
pycparser==2.21
tomli==2.0.1
torch==1.9.0
torchvision==0.9.1
typed-ast==1.5.4
typing_extensions==4.2.0

# Other requirements
black
mypy-extensions
